+++ crypto/src/main/scala/org/bitcoins/crypto/CryptoContext.scala
@@ -28,8 +28,8 @@ object CryptoContext {
   /** The platform specific cryptographic functions required to run bitcoin-s */
   lazy val cryptoRuntime: CryptoRuntime = {
     default match {
-      case LibSecp256k1 => CryptoUtil
-      case BouncyCastle => CryptoUtil
+      case LibSecp256k1 => JvmCryptoRuntime
+      case BouncyCastle => JvmCryptoRuntime
     }
   }
 }
